backend/requirements.txt
# Django
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11

# Database
psycopg2-binary>=2.9
pgvector>=0.2

# Celery
celery>=5.3
redis>=5.0

# AI
langchain>=0.1
langchain-anthropic>=0.1
anthropic>=0.18
google-generativeai>=0.3

# Document Generation
python-docx>=1.1
openpyxl>=3.1
pandas>=2.2
xlsxwriter>=3.2

# Charts
matplotlib>=3.8
plotly>=5.18

# PDF
weasyprint>=60.0

# Images
Pillow>=10.2

# Utilities
python-dotenv>=1.0
requests>=2.31
